Predict top 12 items based on seasonal popularity

+++ MaximGonnissen/seasonality_analysis.py
@@ -1,8 +1,9 @@
+import datetime
 import time
 import multiprocessing as mp
 
 from season import Season, seasons
-from utils import DataFileNames, load_data_from_hnm, get_data_path
+from utils import DataFileNames, load_data_from_hnm, get_data_path, load_data, ProjectConfig
 from progress_bar import ProgressBar
 
 import pandas as pd
@@ -93,28 +94,87 @@ def calculate_season_scores(df: pd.DataFrame) -> pd.DataFrame:
     return new_df
 
 
+def calculate_top_sales(df: pd.DataFrame, start_date: datetime.datetime, end_date: datetime.datetime, top_x: int = 12) -> pd.DataFrame:
+    """
+    Calculate best-selling seasonal items
+    :param df: Dataframe with season scores per item
+    :param start_date: Date to start from
+    :param end_date: Date to end at
+    :param top_x: Calculate top x best-selling items
+    :return: Dataframe with top x best-selling items for each day
+    """
+    out_df = pd.DataFrame(columns=['date', 'items'])
+    out_df['date'] = pd.to_datetime(out_df['date'])
+
+    for day in pd.date_range(start_date, end_date):
+        season = get_season(day)
+        top_x_items = df.sort_values(by=season.season_name, ascending=False).head(top_x)['article_id'].tolist()
+        out_df = out_df._append({'date': day, 'items': ' '.join([str(item_id) for item_id in top_x_items])}, ignore_index=True)
+
+    return out_df
+
+
+def predict_top_items(df: pd.DataFrame) -> list:
+    """
+    Predict most popular items in the period
+    :param df: Dataframe with top items per day based on season scores
+    :return: List of top items
+    """
+    # Go through each row, items are already sorted based on season scores
+    # Each item gets a score based on the position in the list (leftmost = highest score)
+    # The item with the highest score is the most popular item
+    # Make a final selection of the top X items, where X is the length of the item column in the original dataframe
+
+    item_scores = {}
+    for index, row in df.iterrows():
+        items = row['items'].split(' ')
+        for i in range(len(items)):
+            item_scores[items[i]] = item_scores.get(items[i], 0) + len(items) - i
+
+    return sorted(item_scores, key=item_scores.get, reverse=True)[:len(df['items'].iloc[0].split(' '))]
+
+
 if __name__ == '__main__':
     script_start_time = time.time()
     mp_pool_count = max(mp.cpu_count() - 1, 1)
     print(f'Using {mp_pool_count} cores for multiprocessing.')
 
-    # Load data
-    transactions_train = load_data_from_hnm(DataFileNames.TRANSACTIONS_TRAIN)
+    seasonal_scores_path = get_data_path() / DataFileNames.OUTPUT_DIR / 'seasonal_scores.csv'
 
-    # Calculate sales per date for each article
-    article_sales_per_date_df = article_sales_per_date(transactions_train)
+    if not seasonal_scores_path.exists():
+        # Load data
+        transactions_train = load_data_from_hnm(DataFileNames.TRANSACTIONS_TRAIN, dtype={'article_id': str})
 
-    output_path = get_data_path() / DataFileNames.OUTPUT_DIR / 'article_sales_per_date.csv'
-    article_sales_per_date_df.to_csv(output_path, index=False)
+        # Calculate sales per date for each article
+        article_sales_per_date_df = article_sales_per_date(transactions_train)
 
-    # Calculate seasonal sales numbers for each article
-    seasonal_sales_df = calculate_seasonal_sales(article_sales_per_date_df)
+        output_path = get_data_path() / DataFileNames.OUTPUT_DIR / 'article_sales_per_date.csv'
+        article_sales_per_date_df.to_csv(output_path, index=False)
 
-    output_path = get_data_path() / DataFileNames.OUTPUT_DIR / 'seasonal_sales.csv'
-    seasonal_sales_df.to_csv(output_path, index=False)
+        # Calculate seasonal sales numbers for each article
+        seasonal_sales_df = calculate_seasonal_sales(article_sales_per_date_df)
 
-    # Calculate seasonal scores for each article
-    seasonal_scores_df = calculate_season_scores(article_sales_per_date_df)
+        output_path = get_data_path() / DataFileNames.OUTPUT_DIR / 'seasonal_sales.csv'
+        seasonal_sales_df.to_csv(output_path, index=False)
 
-    output_path = get_data_path() / DataFileNames.OUTPUT_DIR / 'seasonal_scores.csv'
-    seasonal_scores_df.to_csv(output_path, index=False)
+        # Calculate seasonal scores for each article
+        seasonal_scores_df = calculate_season_scores(article_sales_per_date_df)
+
+        output_path = seasonal_scores_path
+        seasonal_scores_df.to_csv(output_path, index=False)
+
+    seasonal_scores_df = load_data(seasonal_scores_path, dtype={'article_id': str})
+
+    top_seasonal_sales = calculate_top_sales(seasonal_scores_df, ProjectConfig.DATA_START, ProjectConfig.DATA_START + datetime.timedelta(days=7))
+
+    output_path = get_data_path() / DataFileNames.OUTPUT_DIR / 'top_seasonal_sales.csv'
+    top_seasonal_sales.to_csv(output_path, index=False)
+
+    top_items = predict_top_items(top_seasonal_sales)
+    top_items_string = ' '.join([str(item_id) for item_id in top_items])
+
+    submission_df = load_data_from_hnm(DataFileNames.SAMPLE_SUBMISSION)
+    submission_df['prediction'] = top_items_string
+
+    output_path = get_data_path() / DataFileNames.OUTPUT_DIR / 'submission.csv'
+    submission_df.to_csv(output_path, index=False)
